#84 - Moved files to proper location
